Moved ObfuscatableValue and SimplePrivacyFilterBanner out
