+++ src/components/TemplateCommon/Privacy/obfuscatableValueStyles.scss
@@ -1,0 +1,5 @@
+@import '../../../styles/variables.scss';
+
+.remove-icon {
+    color: $invalid-color;
+}
